+++ packages/react-components/react-headless-components-preview/library/src/components/Drawer/DrawerHeader/renderDrawerHeader.ts
@@ -1,0 +1,6 @@
+import { renderDrawerHeader_unstable } from '@fluentui/react-drawer';
+
+/**
+ * Renders the final JSX of the DrawerHeader component, given the state.
+ */
+export const renderDrawerHeader = renderDrawerHeader_unstable;