+++ packages/react-components/react-headless-components-preview/stories/src/Breadcrumb/index.stories.tsx
@@ -3,14 +3,13 @@ import {
   BreadcrumbItem,
   BreadcrumbButton,
   BreadcrumbDivider,
-} from '@fluentui/react-headless-components-preview';
+} from '@fluentui/react-headless-components-preview/breadcrumb';
 
 import descriptionMd from './BreadcrumbDescription.md';
-
 export { Default } from './BreadcrumbDefault.stories';
 
 export default {
-  title: 'Headless Components/Breadcrumb',
+  title: 'Components/Breadcrumb',
   component: Breadcrumb,
   subcomponents: { BreadcrumbItem, BreadcrumbButton, BreadcrumbDivider },
   parameters: {