+++ packages/react-components/react-headless-components-preview/stories/src/Dialog/index.stories.tsx
@@ -5,10 +5,9 @@ import {
   DialogSurface,
   DialogTitle,
   DialogTrigger,
-} from '@fluentui/react-headless-components-preview';
+} from '@fluentui/react-headless-components-preview/dialog';
 
 import descriptionMd from './DialogDescription.md';
-
 export { Default } from './DialogDefault.stories';
 export { NonModal } from './DialogNonModal.stories';
 export { Alert } from './DialogAlert.stories';
@@ -19,7 +18,7 @@ export { Nested } from './DialogNested.stories';
 export { NoTrigger } from './DialogNoTrigger.stories';
 
 export default {
-  title: 'Headless Components/Dialog',
+  title: 'Components/Dialog',
   component: Dialog,
   subcomponents: { DialogSurface, DialogTrigger, DialogTitle, DialogBody, DialogActions },
   parameters: {