+++ packages/react-components/react-headless-components-preview/stories/src/Select/index.stories.tsx
@@ -1,11 +1,10 @@
-import { Select } from '@fluentui/react-headless-components-preview';
+import { Select } from '@fluentui/react-headless-components-preview/select';
 
 import descriptionMd from './SelectDescription.md';
-
 export { Default } from './SelectDefault.stories';
 
 export default {
-  title: 'Headless Components/Select',
+  title: 'Components/Select',
   component: Select,
   parameters: {
     docs: {