+++ packages/react-components/react-headless-components-preview/stories/src/Slider/index.stories.tsx
@@ -1,11 +1,10 @@
-import { Slider } from '@fluentui/react-headless-components-preview';
+import { Slider } from '@fluentui/react-headless-components-preview/slider';
 
 import descriptionMd from './SliderDescription.md';
-
 export { Default } from './SliderDefault.stories';
 
 export default {
-  title: 'Headless Components/Slider',
+  title: 'Components/Slider',
   component: Slider,
   parameters: {
     docs: {