+++ packages/react-components/react-headless-components-preview/stories/src/Spinner/index.stories.tsx
@@ -1,12 +1,11 @@
-import { Spinner } from '@fluentui/react-headless-components-preview';
+import { Spinner } from '@fluentui/react-headless-components-preview/spinner';
 
 import descriptionMd from './SpinnerDescription.md';
-
 export { Default } from './SpinnerDefault.stories';
 export { Labels } from './SpinnerLabels.stories';
 
 export default {
-  title: 'Headless Components/Spinner',
+  title: 'Components/Spinner',
   component: Spinner,
   parameters: {
     docs: {