+++ packages/web-components/src/text-input/text-input.options.ts
@@ -1,3 +1,4 @@
+import { FluentDesignSystem } from '../fluent-design-system.js';
 import type { StartEndOptions } from '../patterns/start-end.js';
 import type { ValuesOf } from '../utils/typings.js';
 import type { TextInput } from './text-input.js';
@@ -70,3 +71,10 @@ export const ImplicitSubmissionBlockingTypes = [
   'url',
   'week',
 ];
+
+/**
+ * The tag name for the text input element.
+ *
+ * @public
+ */
+export const tagName = `${FluentDesignSystem.prefix}-text-input` as const;