+++ packages/web-components/src/tree-item/tree-item.template.ts
@@ -10,7 +10,7 @@ const chevronIcon = html`
 `;
 
 export const template = html<TreeItem>`
-  <template slot="${x => (x.isNestedItem ? 'item' : void 0)}">
+  <template slot="${x => (x.isNestedItem ? 'item' : void 0)}" tabindex="0" ?focusgroupstart="${x => x.selected}">
     <div class="positioning-region" part="positioning-region">
       <div class="content" part="content">
         <span class="chevron" part="chevron" aria-hidden="true">